Improve cleanipynb hooks
Extend cleanipynb with an argument that allows leaving output cells
as-is, and pass this argument when processing notebook files that are
immediate descendents of the notebooks/ folder so that example outputs
are retained but metadata are removed.  Another hook runs for notebooks
in the notebooks/helpers/ folder so that both outputs and metadata are
removed.